Made labels in GlobalMenu on levels to colorized circles
